DRONI-169 <feat> : droni catting websocket 설정 [지라 이슈 링크]

+++ src/main/java/droni/backend/api/droniuser/repository/DroniUserRepository.java
@@ -3,5 +3,5 @@ package droni.backend.api.droniuser.repository;
 import droni.backend.api.droniuser.entity.DroniUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DroniUserRepository extends JpaRepository<DroniUser, Integer> , DroniUserQuerydslRepository{
+public interface DroniUserRepository extends JpaRepository<DroniUser, Long>, DroniUserQuerydslRepository {
 }
